requirements.txt
fastapi==0.114.2
SQLAlchemy==2.0.34
# psycopg2-binary==2.9.9
pydantic==2.9.1
python-dotenv==1.0.1
uvicorn==0.30.6
alembic==1.13.2
isort==5.13.2
flake8==7.1.1
greenlet==3.1.1
asyncpg==0.29.0